Added alternative Cartesian Grids (untested)
Added new Cartesian grids which store data on different cell
locations. Includes edge centred, face centred, and vertex
centred.

Will now need to move functions around, since start() and
end() functions depend on which type of grid is in use.

A simpler solution would be to implement Face, Edge and
Vertex cented grids such that they all have an extra buffer
cell in each dimension, and conventions are applied to ensure
they behave like the right type of grid, i.e. implement all like
vertex centred, but ignore buffer cell in one or more dimension.
This would waste memory, but would simplify things.

+++ src/Grid/CartesianGrid.hpp
@@ -99,7 +99,7 @@ public:
 
     // Get total number of cells in grid.
     unsigned int total_cells() const {
-        double total = 1.0;
+        unsigned int total = 1;
         for( unsigned int s=0; s<dim; s++){
             total *= size(s);
         }
